% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intense_dependency.R
\name{intense_dependency}
\alias{intense_dependency}
\title{Create an HTML dependency for intense-images.js}
\usage{
intense_dependency()
}
\description{
Create an HTML dependency for intense-images.js
}
