% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_intensify.R
\name{img_intensify}
\alias{img_intensify}
\title{Viewing images on full screen}
\usage{
img_intensify()
}
\value{
images that when clicked can be viewed on full screen no matter their initial size
}
\description{
Viewing images on full screen
}
