% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_intensify.R
\name{img_intensify}
\alias{img_intensify}
\title{Viewing images on full screen}
\usage{
img_intensify(target = "img")
}
\arguments{
\item{target}{By default, the function intensifies all the images (<img> tag) however the user can target specific classes}
}
\value{
images that when clicked can be viewed on full screen no matter their initial size
}
\description{
Viewing images on full screen
}
\examples{

if (interactive()) {


ui <- fluidPage(

h1("Click on the image to view it in full screen"),

img(src = "https://www.r-project.org/logo/Rlogo.png",

width = "200px", height = "200px"),

img_intensify()

)

server <- function(input, output) {

}

shinyApp(ui = ui, server = server)

}

}
